test: maximize shared coverage between Qwen3 XML and Coder parsers
Move all generic regression tests (basic extraction, type conversion,
streaming variants, robustness) from the Coder-specific file into the
shared parametrized file so each test runs against both parsers.  Only
behaviour that genuinely differs between the two parsers stays
parser-specific:

- Coder-only: ``streaming_split_tag`` (relies on ``is_tool_call_started``)
  and ``streaming_various_chunk_sizes`` (XML SAX cannot tolerate
  single-character chunks).
- XML-only: ``streaming_missing_opening_tool_call_tag`` (Coder does not
  recover from a missing ``<tool_call>`` opener in streaming mode).

Two assertions were relaxed in the shared file to accept both legitimate
behaviours: content between parallel tool calls (``None`` vs ``"\\n"``)
and the streaming header arguments value (``""`` vs ``"{"``).

Test count rises from 99 to 138 (+39 from cross-parser parametrization)
while ``test_qwen3coder_tool_parser.py`` shrinks from 1260 to 162 lines.
